Delete unnecessary directory
Not even sure why the fuck this was included in with the upload anyway. That's a stupid ass mistake that the IDE made on here...should have definitely been part of the fucking .gitignore. Either way though - deactivated all of the API keys that got leaked (again that's annoying as all fuck) and any other sensitive information that may have been included